chore: rename to useDeviceInformation
